+++ lib/cli/index.js
@@ -27,7 +27,7 @@ import {
   getOriginUrl,
   gitTreeHashes,
   listFiles,
-} from "./envcontext.js";
+} from "./lib/helpers/envcontext.js/index.js";
 import {
   CARGO_CMD,
   CLJ_CMD,
@@ -140,7 +140,7 @@ import {
   readZipEntry,
   recomputeScope,
   splitOutputByGradleProjects,
-} from "./utils.js";
+} from "./lib/helpers/utils.js/index.js";
 let url = import.meta.url;
 if (!url.startsWith("file://")) {
   url = new URL(`file://${import.meta.url}`).toString();
@@ -152,7 +152,15 @@ const selfPJson = JSON.parse(
 );
 const _version = selfPJson.version;
 import { gte, lte } from "semver";
-import { findJSImportsExports } from "./analyzer.js";
+import { findJSImportsExports } from "../helpers/analyzer.js";
+import { collectOSCryptoLibs } from "../helpers/cbomutils.js";
+import {
+  addSkippedSrcFiles,
+  exportArchive,
+  exportImage,
+  getPkgPathList,
+  parseImageName,
+} from "../managers/docker.js";
 import {
   executeOsQuery,
   getBinaryBom,
@@ -160,14 +168,6 @@ import {
   getDotnetSlices,
   getOSPackages,
 } from "./binary.js";
-import { collectOSCryptoLibs } from "./cbomutils.js";
-import {
-  addSkippedSrcFiles,
-  exportArchive,
-  exportImage,
-  getPkgPathList,
-  parseImageName,
-} from "./docker.js";
 
 const isWin = _platform() === "win32";
