+++ lib/enviser/index.js
@@ -4,8 +4,8 @@ import path from "node:path";
 import process from "node:process";
 import { PackageURL } from "packageurl-js";
 import { Op } from "sequelize";
-import { findCryptoAlgos } from "./cbomutils.js";
-import * as db from "./db.js";
+import { findCryptoAlgos } from "../helpers/cbomutils.js";
+import * as db from "../helpers/db.js";
 import {
   DEBUG_MODE,
   collectGradleDependencies,
@@ -15,7 +15,7 @@ import {
   getGradleCommand,
   getMavenCommand,
   getTimestamp,
-} from "./utils.js";
+} from "../helpers/utils.js";
 const DB_NAME = "evinser.db";
 const typePurlsCache = {};
