+++ lib/enviser/index.test.js
@@ -4,13 +4,15 @@ import {
   constructServiceName,
   detectServicesFromUsages,
   extractEndpoints,
-} from "./evinser.js";
+} from "./index.js";
 
 import { readFileSync } from "node:fs";
 
 test("Service detection test", () => {
   const usageSlice = JSON.parse(
-    readFileSync("./test/data/usages.json", { encoding: "utf-8" }),
+    readFileSync("./../../test/data/usages.json", {
+      encoding: "utf-8",
+    }),
   );
   const objectSlices = usageSlice.objectSlices;
   const servicesMap = {};
