+++ lib/helpers/analyzer.js
