+++ lib/helpers/cbomutils.js
@@ -1,12 +1,15 @@
 import { readFileSync } from "node:fs";
 import { join } from "node:path";
-import { executeOsQuery } from "./binary.js";
+import { executeOsQuery } from "../managers/binary.js";
 import { convertOSQueryResults, dirNameStr } from "./utils.js";
 const cbomosDbQueries = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "cbomosdb-queries.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "cbomosdb-queries.json"),
+    "utf-8",
+  ),
 );
 const cbomCryptoOids = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "crypto-oid.json"), "utf-8"),
+  readFileSync(join(dirNameStr, "../../", "data", "crypto-oid.json"), "utf-8"),
 );
 
 /**
