+++ lib/helpers/cbomutils.test.js
