+++ lib/helpers/db.js
