+++ lib/helpers/display.js
