+++ lib/helpers/display.test.js
