+++ lib/helpers/protobom.js
