+++ lib/helpers/utils.js
@@ -49,7 +49,7 @@ import {
 import toml from "toml";
 import { IriValidationStrategy, validateIri } from "validate-iri";
 import { xml2js } from "xml-js";
-import { getTreeWithPlugin } from "./piptree.js";
+import { getTreeWithPlugin } from "../managers/piptree.js";
 
 let url = import.meta.url;
 if (!url.startsWith("file://")) {
@@ -66,40 +66,64 @@ if (isWin) {
 }
 
 const licenseMapping = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "lic-mapping.json"), "utf-8"),
+  readFileSync(join(dirNameStr, "../../", "data", "lic-mapping.json"), "utf-8"),
 );
 const vendorAliases = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "vendor-alias.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "vendor-alias.json"),
+    "utf-8",
+  ),
 );
 const spdxLicenses = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "spdx-licenses.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "spdx-licenses.json"),
+    "utf-8",
+  ),
 );
 const knownLicenses = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "known-licenses.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "known-licenses.json"),
+    "utf-8",
+  ),
 );
 const mesonWrapDB = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "wrapdb-releases.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "wrapdb-releases.json"),
+    "utf-8",
+  ),
 );
 export const frameworksList = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "frameworks-list.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "frameworks-list.json"),
+    "utf-8",
+  ),
 );
 const selfPJson = JSON.parse(
-  readFileSync(join(dirNameStr, "package.json"), "utf-8"),
+  readFileSync(join(dirNameStr, "../../", "package.json"), "utf-8"),
 );
 
 const CPP_STD_MODULES = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "glibc-stdlib.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "glibc-stdlib.json"),
+    "utf-8",
+  ),
 );
 const _version = selfPJson.version;
 
 // Refer to contrib/py-modules.py for a script to generate this list
 // The script needs to be used once every few months to update this list
 const PYTHON_STD_MODULES = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "python-stdlib.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "python-stdlib.json"),
+    "utf-8",
+  ),
 );
 // Mapping between modules and package names
 const PYPI_MODULE_PACKAGE_MAPPING = JSON.parse(
-  readFileSync(join(dirNameStr, "data", "pypi-pkg-aliases.json"), "utf-8"),
+  readFileSync(
+    join(dirNameStr, "../../", "data", "pypi-pkg-aliases.json"),
+    "utf-8",
+  ),
 );
 
 // Debug mode flag
