+++ lib/helpers/validator.js
