+++ lib/managers/binary.js
@@ -18,7 +18,7 @@ import {
   adjustLicenseInformation,
   findLicenseId,
   isSpdxLicenseExpression,
-} from "./utils.js";
+} from "../helpers/utils.js";
 
 import { URL, fileURLToPath } from "node:url";
