+++ lib/managers/docker.js
@@ -24,7 +24,7 @@ import { parse } from "node:url";
 import { globSync } from "glob";
 import got from "got";
 import { x } from "tar";
-import { DEBUG_MODE, getAllFiles } from "./utils.js";
+import { DEBUG_MODE, getAllFiles } from "../helpers/utils.js";
 
 export const isWin = _platform() === "win32";
 export const DOCKER_HUB_REGISTRY = "docker.io";
