+++ lib/managers/docker.test.js
