+++ lib/managers/piptree.js
