+++ lib/server/index.js
@@ -7,7 +7,7 @@ import process from "node:process";
 import url from "node:url";
 import bodyParser from "body-parser";
 import connect from "connect";
-import { createBom, submitBom } from "./index.js";
+import { createBom, submitBom } from "../cli/index.js";
 import { postProcess } from "./postgen.js";
 
 import compression from "compression";
