+++ lib/stages/postgen/index.test.js
@@ -1,4 +1,4 @@
-import { filterBom } from "./postgen.js";
+import { filterBom } from "./index.js";
 
 import { readFileSync } from "node:fs";
 import { expect, test } from "@jest/globals";
