+++ lib/stages/pregen/index.js
@@ -9,8 +9,8 @@ import {
   installSdkmanTool,
   isNvmAvailable,
   isSdkmanAvailable,
-} from "./envcontext.js";
-import { DEBUG_MODE, hasAnyProjectType } from "./utils.js";
+} from "../../helpers/envcontext.js/index.js";
+import { DEBUG_MODE, hasAnyProjectType } from "../../helpers/utils.js/index.js";
 
 /**
  * Method to prepare the build environment for BOM generation purposes.
