+++ package.json
@@ -40,15 +40,15 @@
   "exports": "./index.js",
   "types": "./types/index.d.ts",
   "bin": {
-    "cdxgen": "bin/cdxgen.js",
-    "obom": "bin/cdxgen.js",
-    "cbom": "bin/cdxgen.js",
-    "cdxi": "bin/repl.js",
-    "evinse": "bin/evinse.js",
-    "cdx-verify": "bin/verify.js"
+    "cdxgen": "./bin/cdxgen.js",
+    "obom": "./bin/cdxgen.js",
+    "cbom": "./bin/cdxgen.js",
+    "cdxi": "./bin/repl.js",
+    "evinse": "./bin/evinse.js",
+    "cdx-verify": "./bin/verify.js"
   },
   "scripts": {
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --inject-globals false docker.test.js utils.test.js display.test.js postgen.test.js",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --inject-globals false ./lib/**/*.test.js",
     "watch": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watch --inject-globals false",
     "lint:check": "biome check",
     "lint": "biome check --fix",
@@ -115,7 +115,7 @@
     "sequelize": "^6.37.3",
     "sqlite3": "^5.1.7"
   },
-  "files": ["*.js", "bin/", "data/", "types/"],
+  "files": ["bin/**/*.js", "lib/**/*.js", "data/", "types/"],
   "devDependencies": {
     "@biomejs/biome": "1.8.3",
     "jest": "^29.7.0",
